+++ src/middleware/services/tti/index.ts
@@ -1,2 +1,3 @@
 export * from './tti.service';
 export * from './providers';
+export * from './utils';